requirements.txt
azure-functions
azure-functions-durable
azure-identity
azure-servicebus
azure-cosmos
sqlalchemy
pyodbc
openai
